CLANG_WARN_QUOTED_INCLUDE_IN_FRAMEWORK_HEADER = YES_ERROR
===== CLANG_WARN_QUOTED_INCLUDE_IN_FRAMEWORK_HEADER = YES_ERROR =====

This warning is an Xcode 12 recommendation (technically, the warning
is not YES_ERROR, but setting error will ensure that future changes
won't regress consumers of OCMock that also set this setting).

This setting will help assure modules importing OCMock that it's
closure of headers is all modular, and will allow projects also
setting this setting to build against OCMock without error.

Inspection & testingreveals it is also available already for Xcode11.

Changes to .h are the result of
- setting CLANG_WARN_QUOTED_INCLUDE_IN_FRAMEWORK_HEADER = YES_ERROR
- building
- applying the fix recommended by Xcode