Fixed the logic for retaining/releasing the targets.
I believe that for almost all cases this code does the same as before (all tests pass unchanged) but it addresses the issue that the init target retain did, in fact, not balance the release in the stub. Unfortunately, I couldn't think of a test that proves that this code is "more correct".